Moved createIssueFactory to helpers
