+++ src/main/core.ts
@@ -1,6 +1,5 @@
 export * from './dsl';
 export * from './shapes';
 export * from './types';
-export { createIssueFactory } from './utils';
 export { Type, TypeName } from './Type';
 export { ValidationError } from './ValidationError';
