+++ src/main/shapes/EnumShape.ts
@@ -1,7 +1,8 @@
 import { CODE_ENUM, MESSAGE_ENUM } from '../constants';
+import { createIssueFactory } from '../helpers';
 import { TYPE_ARRAY, TYPE_OBJECT } from '../Type';
 import { ApplyOptions, ConstraintOptions, Message } from '../types';
-import { createIssueFactory, getCanonicalValueOf, isArray, ok, ReadonlyDict, unique } from '../utils';
+import { getCanonicalValueOf, isArray, ok, ReadonlyDict, unique } from '../utils';
 import { CoercibleShape } from './CoercibleShape';
 import { NEVER, Result } from './Shape';
