+++ src/test/utils/shapes.test.ts
@@ -1,5 +1,5 @@
 import { Issue, Shape, ValidationError } from '../../main';
-import { copyUnsafeChecks, createApplyChecksCallback, createIssueFactory } from '../../main/utils';
+import { copyUnsafeChecks, createApplyChecksCallback } from '../../main/utils';
 
 describe('copyUnsafeChecks', () => {
   test('returns the exact copy of the target shape is there are no unsafe check on the source shape', () => {
@@ -21,90 +21,6 @@ describe('copyUnsafeChecks', () => {
     expect(shape).not.toBe(targetShape);
     expect(shape['_checks'].length).toBe(1);
     expect(shape['_checks'][0].callback).toBe(unsafeCheck);
-  });
-});
-
-describe('createIssueFactory', () => {
-  describe('known param', () => {
-    test('creates a factory with the default message', () => {
-      const issueFactory = createIssueFactory('aaa', 'bbb', undefined, 'eee');
-
-      expect(issueFactory('xxx', {})).toEqual([{ code: 'aaa', input: 'xxx', message: 'bbb', param: 'eee' }]);
-    });
-
-    test('creates a factory with a string message', () => {
-      const issueFactory = createIssueFactory('aaa', 'bbb', 'ccc %s', 'eee');
-
-      expect(issueFactory('xxx', {})).toEqual([{ code: 'aaa', input: 'xxx', message: 'ccc eee', param: 'eee' }]);
-    });
-
-    test('creates a factory with a function message', () => {
-      const cbMock = jest.fn(() => 222);
-      const issueFactory = createIssueFactory('aaa', 'bbb', cbMock, 'eee');
-
-      expect(issueFactory('xxx', {})).toEqual([{ code: 'aaa', input: 'xxx', message: 222, param: 'eee' }]);
-      expect(cbMock).toHaveBeenCalledTimes(1);
-      expect(cbMock).toHaveBeenNthCalledWith(1, 'eee', 'aaa', 'xxx', undefined, {});
-    });
-
-    test('creates a factory with a string message in options', () => {
-      const issueFactory = createIssueFactory('aaa', 'bbb', { message: 'ccc %s', meta: 111 }, 'eee');
-
-      expect(issueFactory('xxx', {})).toEqual([
-        { code: 'aaa', input: 'xxx', message: 'ccc eee', meta: 111, param: 'eee' },
-      ]);
-    });
-
-    test('creates a factory with a function message in options', () => {
-      const cbMock = jest.fn(() => 222);
-      const issueFactory = createIssueFactory('aaa', 'bbb', { message: cbMock, meta: 111 }, 'eee');
-
-      expect(issueFactory('xxx', {})).toEqual([{ code: 'aaa', input: 'xxx', message: 222, meta: 111, param: 'eee' }]);
-      expect(cbMock).toHaveBeenCalledTimes(1);
-      expect(cbMock).toHaveBeenNthCalledWith(1, 'eee', 'aaa', 'xxx', 111, {});
-    });
-  });
-
-  describe('unknown param', () => {
-    test('creates a factory with the default message', () => {
-      const issueFactory = createIssueFactory('aaa', 'bbb', undefined);
-
-      expect(issueFactory('xxx', {}, 'eee')).toEqual([{ code: 'aaa', input: 'xxx', message: 'bbb', param: 'eee' }]);
-    });
-
-    test('creates a factory with a string message', () => {
-      const issueFactory = createIssueFactory('aaa', 'bbb', 'ccc %s');
-
-      expect(issueFactory('xxx', {}, 'eee')).toEqual([{ code: 'aaa', input: 'xxx', message: 'ccc eee', param: 'eee' }]);
-    });
-
-    test('creates a factory with a function message', () => {
-      const cbMock = jest.fn(() => 222);
-      const issueFactory = createIssueFactory('aaa', 'bbb', cbMock);
-
-      expect(issueFactory('xxx', {}, 'eee')).toEqual([{ code: 'aaa', input: 'xxx', message: 222, param: 'eee' }]);
-      expect(cbMock).toHaveBeenCalledTimes(1);
-      expect(cbMock).toHaveBeenNthCalledWith(1, 'eee', 'aaa', 'xxx', undefined, {});
-    });
-
-    test('creates a factory with a string message in options', () => {
-      const issueFactory = createIssueFactory('aaa', 'bbb', { message: 'ccc %s', meta: 111 });
-
-      expect(issueFactory('xxx', {}, 'eee')).toEqual([
-        { code: 'aaa', input: 'xxx', message: 'ccc eee', meta: 111, param: 'eee' },
-      ]);
-    });
-
-    test('creates a factory with a function message in options', () => {
-      const cbMock = jest.fn(() => 222);
-      const issueFactory = createIssueFactory('aaa', 'bbb', { message: cbMock, meta: 111 });
-
-      expect(issueFactory('xxx', { context: 333 }, 'eee')).toEqual([
-        { code: 'aaa', input: 'xxx', message: 222, meta: 111, param: 'eee' },
-      ]);
-      expect(cbMock).toHaveBeenCalledTimes(1);
-      expect(cbMock).toHaveBeenNthCalledWith(1, 'eee', 'aaa', 'xxx', 111, { context: 333 });
-    });
   });
 });
